scripts/explainability.py
import shap
import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestRegressor  # Example model (replace with your actual model)

# Example data (use your actual match scores and features)
# For demonstration purposes, I will use mock data
data = pd.DataFrame({
    'skills_score': [0.9, 0.5, 0.75, 0.8],
    'experience_score': [0.6, 0.3, 0.7, 0.4],
    'education_score': [0.7, 0.8, 0.9, 0.6],
    'final_score': [0.85, 0.6, 0.8, 0.75]  # Target variable
})

# Define your features and target (final score)
X = data[['skills_score', 'experience_score', 'education_score']]  # Features
y = data['final_score']  # Target variable (final score)

# Fit a model (RandomForestRegressor in this case, but you can use your own model)
model = RandomForestRegressor()
model.fit(X, y)

# Create a SHAP explainer object
explainer = shap.TreeExplainer(model)  # Use appropriate explainer based on your model type
shap_values = explainer.shap_values(X)

# Visualize SHAP values (this will open in a browser or display inline depending on your environment)
shap.summary_plot(shap_values, X)

# You can also save SHAP values or visualizations
shap.initjs()
shap.save_html("shap_summary.html", shap.summary_plot(shap_values, X, plot_type="bar"))
