requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
fsspec==2025.3.2
gunicorn==23.0.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lime==0.2.0.1
llvmlite==0.44.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.36.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pdfminer.six==20250416
pillow==11.2.1
plotly==6.0.1
pluggy==1.5.0
preshed==3.0.9
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.1.0
shap==0.47.2
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
